fixed build to only use my linker script, not mine and the default
Added the -T parameter in front of the linker script name in the Makefile
which causes the linker to use ONLY this script.  This was required to get
the executable to work properly under Debian based distributions but was
causing things to puke on my Gentoo system.  Fixing this required one very
slight change to the linker script itself

Updated the contents of the README directory - this really needs a proper
going over some day :)

Have noticed a difference between the elf executable produced by the nasm
assembler and the one produced by my fsave extension in that readelf -a
shows the dummy section header zero on the kernel but not on the extened
forth.  Have made some tweaks to fsave.f trying to figure out why.  This
does not seem to be a catastrophic issue but it bugs my OCD :)